Clean up wal
