+++ src/engine.rs
@@ -1,6 +1,3 @@
-use std::fmt::Debug;
-
-use log::*;
 /// The engine is the main abstraction to interact with the key-value store system.
 ///
 /// The engine owns all key-value pairs and takes care of
@@ -12,10 +9,11 @@ use log::*;
 /// It presents itself with a dictionary-like interface where each operation
 /// might fail. This is deliberate since every operation has to potentially interact
 /// with the OS or the network which are unreliable components.
-use thiserror::Error;
-
 // re-exports for convenience
 pub use key::Key;
+use log;
+use std::fmt::Debug;
+use thiserror::Error;
 pub use value::Value;
 
 use crate::engine::configuration::Configuration;
@@ -41,13 +39,16 @@ pub enum Error {
     FileSystemError(#[from] directories::Error),
 }
 
+/// The engine encapsulates local storage, replication and distribution
 pub struct Engine {
     lsm: storage::lsm::LSM,
 }
 
 impl Engine {
-    pub fn new(config: Configuration) -> Result<Self> {
+    /// Starts up the engine and makes sure it's operational.
+    pub fn start(config: Configuration) -> Result<Self> {
         directories::setup_directories(&config)?;
+
         let lsm = storage::lsm::LSM::new(config.storage)?;
         Ok(Self { lsm })
     }
@@ -62,7 +63,7 @@ impl Engine {
         key: K,
         value: V,
     ) -> Result<Option<Value>> {
-        trace!(target: "engine", "Insert {:?} -> {:?}", key, value);
+        log::trace!(target: "engine", "Insert {:?} -> {:?}", key, value);
         Ok(self.lsm.set(key.into(), value.into())?)
     }
 
@@ -76,7 +77,7 @@ impl Engine {
     /// * the change is durable on the local node.
     /// * the key/value can not be found anymore (unless it has been re-inserted)
     pub fn del(&mut self, key: &Key) -> Result<Option<Value>> {
-        trace!(target: "engine", "Delete {:?}", key);
+        log::trace!(target: "engine", "Delete {:?}", key);
         Ok(self.lsm.del(&key)?)
     }
 
@@ -86,7 +87,7 @@ impl Engine {
     /// This operation might fail, e.g. when implementatons need to access the
     /// filesystem or the network.
     pub fn get(&self, key: &Key) -> Result<Option<&Value>> {
-        trace!(target: "engine", "Lookup {:?}", key);
+        log::trace!(target: "engine", "Lookup {:?}", key);
         Ok(self.lsm.get(&key)?)
     }
